website/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.12",
    "onnxruntime>=1.20.1",
    "opencv-python-headless>=4.11.0.86",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.4",
    "python-jose[cryptography]>=3.4.0",
    "sqlalchemy>=2.0.40",
    "ultralytics>=8.3.140",
    "uvicorn>=0.34.2",
]
